allow erlmld_worker:ready/1 callback to checkpoint
- `erlmld_worker:ready/1` behavior callback may now checkpoint (same
  as `erlmld_worker:process_record/2`)

- optional `erlmld_flusher:ready/1` behavior callback added which may
  return same result as `erlmld_flusher:flush/2`.  if it returns any
  tokens, treated as if a flush occurred.

With these changes, when using an `erlmld_flusher`-implementing module
with `erlmld_batch_processor`, it should no longer be possible to be
permanently stuck in a state where `erlmld_batch_processor` is waiting
on the flusher, which is itself waiting on more events from the batch
processor before flushing anything (such as might happen when
processing a low-volume stream).

Now, if a flusher module exports `ready/1`, it will be called
periodically (and before each record batch) even if no records are
available on the stream, giving it an opportunity to flush even if its
batch is not full.